Relocate command classes
